+++ farm_regen_digital.post_update.php
@@ -22,3 +22,12 @@ function farm_regen_digital_post_update_enable_farm_import_kml(&$sandbox = NULL)
     \Drupal::service('module_installer')->install(['farm_import_kml']);
   }
 }
+
+/**
+ * Install farm_regen_digital_map module.
+ */
+function farm_regen_digital_post_update_enable_farm_regen_digital_map(&$sandbox = NULL) {
+  if (!\Drupal::service('module_handler')->moduleExists('farm_regen_digital_map')) {
+    \Drupal::service('module_installer')->install(['farm_regen_digital_map']);
+  }
+}