+++ modules/map/src/EventSubscriber/MapRenderEventSubscriber.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Drupal\farm_regen_digital_map\EventSubscriber;
+
+use Drupal\farm_map\Event\MapRenderEvent;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+
+/**
+ * An event subscriber for the MapRenderEvent.
+ *
+ * This adds the regen_digital_standard behavior to all maps.
+ */
+class MapRenderEventSubscriber implements EventSubscriberInterface {
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function getSubscribedEvents() {
+    return [
+      MapRenderEvent::EVENT_NAME => 'onMapRender',
+    ];
+  }
+
+  /**
+   * React to the MapRenderEvent.
+   *
+   * @param \Drupal\farm_map\Event\MapRenderEvent $event
+   *   The MapRenderEvent.
+   *
+   * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
+   * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
+   */
+  public function onMapRender(MapRenderEvent $event) {
+    $event->addBehavior('regen_digital_standard');
+  }
+
+}